requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pytest==8.2.2
httpx==0.27.0
SQLAlchemy==2.0.30
alembic==1.13.1
psycopg[binary]==3.2.13
passlib==1.7.4
python-jose==3.3.0

bcrypt==4.1.3
